refactor(fabric): extract logic for contract reads to util function
